code/src/requirements.txt
streamlit
transformers
torch
pandas
numpy
scikit-learn
sentence-transformers
transformers
torch
python-dotenv
plotly
joblib